+++ src/actions/generate.rs
@@ -9,7 +9,7 @@ use crate::{
     log,
     odra_toml::Contract,
     paths::{to_camel_case, to_snake_case},
-    project::Project,
+    project::{OdraLocation, Project},
     template::TemplateGenerator,
 };
 
@@ -20,16 +20,29 @@ pub struct GenerateAction<'a> {
     contract_module_ident: String,
     module_root: PathBuf,
     module_name: Option<String>,
+    template_name: Option<String>,
     template_generator: TemplateGenerator,
 }
 
 /// GenerateAction implementation.
 impl<'a> GenerateAction<'a> {
     /// Crate a new GenerateAction for a given contract.
-    pub fn new(project: &'a Project, contract_name: String, module_name: Option<String>) -> Self {
+    pub fn new(
+        project: &'a Project,
+        contract_name: String,
+        module_name: Option<String>,
+        template_name: Option<String>,
+        source: Option<String>,
+    ) -> Self {
         if project.is_workspace() && module_name.is_none() {
             Error::CrateNotProvided.print_and_die();
         }
+
+        let odra_location = match source {
+            None => project.project_odra_location(),
+
+            Some(_) => OdraLocation::from_source(source),
+        };
 
         GenerateAction {
             project,
@@ -37,9 +50,10 @@ impl<'a> GenerateAction<'a> {
             contract_module_ident: to_snake_case(contract_name),
             module_root: project.module_root(module_name.clone()),
             module_name,
+            template_name,
             template_generator: TemplateGenerator::new(
                 ODRA_TEMPLATE_GH_RAW_REPO.to_string(),
-                project.project_odra_location(),
+                odra_location,
             ),
         }
     }
@@ -49,7 +63,7 @@ impl GenerateAction<'_> {
     /// Main function that runs the generation action.
     pub fn generate_contract(&self) {
         log::info(format!("Adding new contract: {} ...", self.contract_name()));
-        self.add_contract_file_to_src();
+        self.add_contract_file_to_src(self.template_name.clone());
         self.update_lib_rs();
         self.update_odra_toml();
     }
@@ -84,11 +98,11 @@ impl GenerateAction<'_> {
     }
 
     /// Crates a new module file in src directory.
-    fn add_contract_file_to_src(&self) {
+    fn add_contract_file_to_src(&self, template_name: Option<String>) {
         // Rename module name.
         let contract_body = self
             .template_generator
-            .module_template(&self.contract_struct_name())
+            .module_template(&self.contract_struct_name(), template_name)
             .unwrap_or_else(|err| err.print_and_die());
 
         // Make sure the file do not exist.